WIP: rename files and refactor throughout
